// frontend/src/company.d.ts
export interface CompanySearch {
  symbol: string;
  name: string;
  currency: string;
  exchangeFullName: string;
  exchange: string;
}